[IMP] event_sale: set proper invoicing policy for event product(s)
Right now, the default invoice policy policy for the products is set to
'delivery'. However, in the case of event ticket products, usually you
don't really handle the "delivery" of the ticket. It's most likely to be
sent by email automatically.

With this commit, the invocing policy for the default ticket product is
set to 'order' to ease inovicing the SO. Apart from default product, the
onchage method for 'event_ok' is adapted for setting the correct invoice
policy and product type.

Task Id : 2243915

closes odoo/odoo#63725

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>