[FIX] base: test failing because of environment reset
The call to reset() when simulating requests during HttpCase tests
leaves the test class into an inconsistent state.  Indeed, the test
environment is no longer into Environment._local, and this may trigger
record cache errors, because new environments use a record cache that is
different from the test's environment.

As this issue only occurs in the context of unit testing, it is not
worth fixing on a stable release, because of the involved risk.
Instead, we modify the test so that the failing operation is made in
method setUpClass(), before any environment reset.

closes odoo/odoo#63730

X-original-commit: 6112e64298be135a5ae93ad5d6c3048ab25fd4b3
Related: odoo/enterprise#15439
Signed-off-by: Raphael Collet (rco) <rco@openerp.com>